// Capstone Project/src/app/pages/user/detail-room-page/screen/types.ts
export enum detailRoomPageEnum {
  GET_DETAIL_ROOM_REQUEST = 'APP/DETAIL_ROOM_PAGE/GET_DETAIL_ROOM_REQUEST',
  GET_DETAIL_ROOM_SUCCESS = 'APP/DETAIL_ROOM_PAGE/GET_DETAIL_ROOM_SUCCESS',
  GET_DETAIL_ROOM_FAILURE = 'APP/DETAIL_ROOM_PAGE/GET_DETAIL_ROOM_FAILURE',

  GET_DETAIL_POST_REQUEST = 'APP/DETAIL_ROOM_PAGE/GET_DETAIL_POST_REQUEST',
  GET_DETAIL_POST_SUCCESS = 'APP/DETAIL_ROOM_PAGE/GET_DETAIL_POST_SUCCESS',
  GET_DETAIL_POST_FAILURE = 'APP/DETAIL_ROOM_PAGE/GET_DETAIL_POST_FAILURE',

  GET_DETAIL_POST_WITH_ROOM_REQUEST = 'APP/GET_DETAIL_POST_WITH_ROOM_REQUEST',
  GET_DETAIL_POST_WITH_ROOM_SUCCESS = 'APP/GET_DETAIL_POST_WITH_ROOM_SUCCESS',
  GET_DETAIL_POST_WITH_ROOM_FAILURE = 'APP/GET_DETAIL_POST_WITH_ROOM_FAILURE',

  REPORT_POST_REQUEST = 'APP/DETAIL_ROOM_PAGE/REPORT_POST_REQUEST',
  REPORT_POST_SUCCESS = 'APP/DETAIL_ROOM_PAGE/REPORT_POST_SUCCESS',
  REPORT_POST_FAILURE = 'APP/DETAIL_ROOM_PAGE/REPORT_POST_FAILURE',


  ADD_TO_FAVOURITE_REQUEST = 'APP/DETAIL_ROOM_PAGE/ADD_TO_FAVOURITE_REQUEST',
  ADD_TO_FAVOURITE_SUCCESS = 'APP/DETAIL_ROOM_PAGE/ADD_TO_FAVOURITE_SUCCESS',
  ADD_TO_FAVOURITE_FAILURE = 'APP/DETAIL_ROOM_PAGE/ADD_TO_FAVOURITE_FAILURE',

  GET_LIST_FAVOURITE_USER_REQUEST = 'APP/DETAIL_ROOM_PAGE/GET_LIST_FAVOURITE_USER_REQUEST',
  GET_LIST_FAVOURITE_USER_SUCCESS = 'APP/DETAIL_ROOM_PAGE/GET_LIST_FAVOURITE_USER_SUCCESS',
  GET_LIST_FAVOURITE_USER_FAILURE = 'APP/DETAIL_ROOM_PAGE/GET_LIST_FAVOURITE_USER_FAILURE',

  GET_LIST_FEEDBACK_REQUEST = 'APP/DETAIL_ROOM_PAGE/GET_LIST_FEEDBACK_REQUEST',
  GET_LIST_FEEDBACK_SUCCESS = 'APP/DETAIL_ROOM_PAGE/GET_LIST_FEEDBACK_SUCCESS',
  GET_LIST_FEEDBACK_FAILURE = 'APP/DETAIL_ROOM_PAGE/GET_LIST_FEEDBACK_FAILURE',

  CREATE_FEEDBACK_REQUEST = 'APP/DETAIL_ROOM_PAGE/CREATE_FEEDBACK_REQUEST',
  CREATE_FEEDBACK_SUCCESS = 'APP/DETAIL_ROOM_PAGE/CREATE_FEEDBACK_SUCCESS',
  CREATE_FEEDBACK_FAILURE = 'APP/DETAIL_ROOM_PAGE/CREATE_FEEDBACK_FAILURE',

  UPDATE_FEEDBACK_REQUEST = 'APP/DETAIL_ROOM_PAGE/UPDATE_FEEDBACK_REQUEST',
  UPDATE_FEEDBACK_SUCCESS = 'APP/DETAIL_ROOM_PAGE/UPDATE_FEEDBACK_SUCCESS',
  UPDATE_FEEDBACK_FAILURE = 'APP/DETAIL_ROOM_PAGE/UPDATE_FEEDBACK_FAILURE',

  CLEAR_STATE_REPORT = 'APP/DETAIL_ROOM_PAGE/CLEAR_STATE_REPORT',

  CLEAR_DETAIL_ROOM_PAGE_STATE = 'APP/DETAIL_ROOM_PAGE/CLEAR_DETAIL_ROOM_PAGE_STATE',
}